+++ src/Controller/ActivityPub/User/UserOutboxController.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Controller\ActivityPub\User;
 
 use App\Controller\AbstractController;
+use App\Entity\Contracts\ActivityPubActivityInterface;
 use App\Entity\User;
-use App\Factory\ActivityPub\PostNoteFactory;
 use App\Repository\UserRepository;
 use App\Service\ActivityPub\Wrapper\CollectionInfoWrapper;
 use App\Service\ActivityPub\Wrapper\CollectionItemsWrapper;
@@ -85,7 +85,7 @@ class UserOutboxController extends AbstractController
             $activity,
             $items,
             $page,
-            PostNoteFactory::ADDITIONAL_CONTEXTS
+            ActivityPubActivityInterface::ADDITIONAL_CONTEXTS
         );
     }
 }
