+++ src/Factory/ActivityPub/PostCommentNoteFactory.php
@@ -38,7 +38,7 @@ class PostCommentNoteFactory
             $note['@context'] = [
                 ActivityPubActivityInterface::CONTEXT_URL,
                 ActivityPubActivityInterface::SECURITY_URL,
-                PostNoteFactory::ADDITIONAL_CONTEXTS,
+                ActivityPubActivityInterface::ADDITIONAL_CONTEXTS,
             ];
         }
